Fix incorrect string conversions in add_chunk() call

+++ modules/dmrpp_module/DMZ.cc
@@ -995,7 +995,7 @@ void DMZ::process_chunk(DmrppCommon *dc, const xml_node &chunk) const
         //  a hash_map of known URLs and cut down on the total number of shared pointers.
         //  jhrg 11/22/21
         shared_ptr<http::url> data_url(new http::url(href, href_trusted));
-        dc->add_chunk(data_url, dc->get_byte_order(), stoi(size), stoi(offset), chunk_position_in_array);
+        dc->add_chunk(data_url, dc->get_byte_order(), stol(size), stol(offset), chunk_position_in_array);
     }
     else {
         dc->add_chunk(d_dataset_elem_href, dc->get_byte_order(), stol(size), stol(offset), chunk_position_in_array);
